# docs/swagger.yaml
definitions:
  models.Cashier:
    properties:
      id:
        type: integer
      serving:
        type: string
    type: object
  models.LoginReq:
    properties:
      password:
        type: string
      user:
        type: string
    type: object
  models.MacReq:
    properties:
      action:
        type: string
      value:
        type: string
    type: object
  models.OrderReq:
    properties:
      ordernum:
        type: string
      ordertime:
        type: string
    type: object
  models.VerifyMacReq:
    properties:
      action:
        type: string
      mac:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login a user
      parameters:
      - description: Login Data
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Login
      tags:
      - user
  /auth/logout:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
      summary: Logout
      tags:
      - user
  /auth/mac:
    post:
      consumes:
      - application/json
      description: Calculate HMAC
      parameters:
      - description: cashier data
        in: body
        name: create
        required: true
        schema:
          $ref: '#/definitions/models.MacReq'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Calculate HMAC for message
      tags:
      - Auth
  /auth/status:
    get:
      consumes:
      - application/json
      description: Check login status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Auth Status
      tags:
      - user
  /auth/verify:
    post:
      consumes:
      - application/json
      description: Verify HMAC
      parameters:
      - description: data
        in: body
        name: create
        required: true
        schema:
          $ref: '#/definitions/models.VerifyMacReq'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Verify MAC for message
      tags:
      - Auth
  /cashiers:
    get:
      consumes:
      - application/json
      description: Get Cashiers
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get Cashiers
      tags:
      - Cashiers
    post:
      consumes:
      - application/json
      description: Create a new Cashier
      parameters:
      - description: cashier data
        in: body
        name: create
        required: true
        schema:
          $ref: '#/definitions/models.Cashier'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create Cashier
      tags:
      - Cashiers
  /cashiers/{cid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Cashier ID
        in: path
        name: cid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: delete cashier
      tags:
      - Cashiers
    patch:
      consumes:
      - application/json
      parameters:
      - description: Cashier ID
        in: path
        name: cid
        required: true
        type: string
      - description: Order Info
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.OrderReq'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: set the order number a cashier is servicing
      tags:
      - Cashiers
  /health:
    get:
      consumes:
      - application/json
      description: Health check
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Health Check
      tags:
      - status
swagger: "2.0"